orchestrator/pyproject.toml
[tool.poetry]
name = "forgefleet-orchestrator"
version = "0.1.0"
description = "ForgeFleet Orchestrator service"
authors = ["You"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"

# Core dependencies
openai-agents = "*"
fastapi = "*"
uvicorn = { extras = ["standard"], version = "*" }
redis = "*"
pydantic = "*"

# PostgreSQL and pgvector
psycopg2-binary = "*"
pgvector = "*"

# gRPC toolchain
grpcio = "1.72.0"
grpcio-tools = "1.72.0"
protobuf = ">=6.30.0"
